Looks made better

+++ TechTask/TechTask/Task2/views.py
@@ -73,7 +73,7 @@ def ViewEvents(request):
     eventList=list()
     for e in EventDet.objects.all():
         eventList.append(model_to_dict(e))
-    return render(request,"eve.html",{'view':1,'eventList':enumerate(eventList)})
+    return render(request,"eve.html",{'view':1,'eventList':enumerate(eventList,1)})
 
 def DeleteEvents(request):
     choice=request.POST.get("CRUDchoice")
@@ -84,10 +84,11 @@ def DeleteEvents(request):
         eventList.append(model_to_dict(e))
     
     delIndex=request.POST.get("DelChoice")
-    for ind,n in enumerate(eventList):
-        if(ind==int(delIndex)):
-            EventDet.objects.filter(name=n['name']).delete()
-            return render(request,"eve.html",{'delete':1,'eventList':enumerate(eventList),'deleted':1})
+    if(delIndex):
+        for ind,n in enumerate(eventList):
+            if(ind==int(delIndex)):
+                EventDet.objects.filter(name=n['name']).delete()
+                return render(request,"eve.html",{'delete':1,'eventList':enumerate(eventList),'deleted':1})
     return render(request,"eve.html",{'delete':1,'eventList':enumerate(eventList)})
 
 def DeleteAll(request):
